Refactor feature:home
