+++ gradle/libs.versions.toml
@@ -20,7 +20,7 @@ dagger-hilt = "2.44.2"
 hilt-navigation-compose = "1.0.0"
 coil = "2.2.2"
 accompanist = "0.27.0"
-date-time-picker = "1.0.2"
+date-time-picker = "1.1.1"
 message-bar-compose = "1.0.5"
 one-tap-compose = "1.0.0"
 desugar = "1.2.0"
